// packages/shared-services/services/models/BankAccountValidationPayload.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type BankAccountValidationPayload = {
    accountNumber?: string | null;
    bankCode?: string | null;
    accountName?: string | null;
    bvn?: string | null;
    phoneNumber?: string | null;
    email?: string | null;
    country?: string | null;
};